world - More ABI breakage
* Make more structural changes that will break ABIs.  Since we are
  breaking ABI's we might as well get as much of it done as possible.

  struct datum		(ndbm and rpcsvc)
  struct stat		(see note below)
  struct ipc_perm	(sysv messaging and ipc)

* The struct stat changes use a spare field so the structure
  size has NOT changed.  The kernel has been modified to fill
  in the 'old' field for ABI compatibility.

  The other structures, however, will break ABIs, particularly
  struct ipc_perm.

* Tested with a full world + kernel build.  Additional work in
  dports will be needed, certainly a whole new package set for
  master (also needed due to other ABI-breaking commits).

Submitted-by: swildner

+++ include/ndbm.h
@@ -48,14 +48,12 @@
 /*
  * The db(3) support for ndbm always appends this suffix to the
  * file name to avoid overwriting the user's original database.
- *
- * XXX dsize should be size_t; see 1003.1-2008
  */
 #define	DBM_SUFFIX	".db"
 
 typedef struct {
 	void *dptr;
-	int dsize;
+	size_t dsize;
 } datum;
 
 typedef DB DBM;
